fix: separate authOptions to avoid error type route.ts nextauth
