+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import { getServerSession } from "next-auth";
-import { authOptions } from "./api/auth/[...nextauth]/route";
 import { redirect } from "next/navigation";
+import authOptions from "./api/auth/[...nextauth]/authOpts";
 
 export default async function Home() {
   const session = await getServerSession(authOptions);
